+++ packages/toolbars/web/src/Toolbar/ButtonComponents/ColorPickerButton.jsx
@@ -34,13 +34,13 @@ class ColorPickerButton extends Component {
     this.setState({ isModalOpen: false });
   };
 
-  onColorAdded = color => {
+  onColorAdded = ({ color }) => {
     this.props.onColorAdded(color);
     const userColors = this.props?.getUserColors?.() || [...this.state.userColors, color];
     this.setState({ userColors });
   };
 
-  onChange = color => {
+  onChange = ({ color }) => {
     this.props.onChange(color);
     this.setState({ currentColor: color });
     this.closeModal();
@@ -53,7 +53,7 @@ class ColorPickerButton extends Component {
       onResetColor();
     } else {
       const defaultColors = getDefaultColors?.();
-      this.onChange(defaultColors);
+      this.onChange({ color: defaultColors });
     }
     this.closeModal();
     this.props.afterClick && this.props.afterClick();